chore: 인수/통합 테스트 및 common(픽스쳐 팩토리, mock api) 패키지 분리  (#217)
