+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/integration/post/PostServiceIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pickgit.post.application;
+package com.woowacourse.pickgit.integration.post;
 
 import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -7,10 +7,12 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.woowacourse.pickgit.authentication.domain.user.GuestUser;
 import com.woowacourse.pickgit.authentication.domain.user.LoginUser;
-import com.woowacourse.pickgit.common.FileFactory;
+import com.woowacourse.pickgit.common.factory.FileFactory;
 import com.woowacourse.pickgit.exception.platform.PlatformHttpErrorException;
 import com.woowacourse.pickgit.exception.post.CommentFormatException;
 import com.woowacourse.pickgit.post.PostTestConfiguration;
+import com.woowacourse.pickgit.common.factory.PostFactory;
+import com.woowacourse.pickgit.post.application.PostService;
 import com.woowacourse.pickgit.post.application.dto.CommentResponse;
 import com.woowacourse.pickgit.post.application.dto.response.PostResponseDto;
 import com.woowacourse.pickgit.post.application.dto.request.PostRequestDto;
