+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/tag/TestTagConfiguration.java
@@ -1,6 +1,6 @@
 package com.woowacourse.pickgit.tag;
 
-import com.woowacourse.pickgit.tag.infrastructure.MockTagApiRequester;
+import com.woowacourse.pickgit.common.mockapi.MockTagApiRequester;
 import com.woowacourse.pickgit.tag.infrastructure.PlatformApiRequester;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
