+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/user/domain/UserRepositoryTest.java
@@ -2,7 +2,7 @@ package com.woowacourse.pickgit.user.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.woowacourse.pickgit.user.UserFactory;
+import com.woowacourse.pickgit.common.factory.UserFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
